Done some changes for contact us page
